backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.12.3
bpy==4.3.0
cachetools==5.3.2
certifi==2025.4.26
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fake-bpy-module-2.80==20250130
fastapi==0.115.12
filelock==3.13.3
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.55.2
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2024.3.1
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.2.1
griffe==1.7.3
grpcio==1.60.1
h11==0.16.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.6
Jinja2==3.1.3
jiter==0.9.0
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.7
libclang==16.0.6
llama-cloud==0.1.19
llama-cloud-services==0.6.22
llama-index==0.12.35
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.35
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-groq==0.3.1
llama-index-llms-openai==0.3.38
llama-index-llms-openai-like==0.3.4
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.22
llvmlite==0.42.0
lxml==5.3.0
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.3
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.59.1
numpy==1.26.3
oauthlib==3.2.2
openai==1.77.0
openai-whisper==20231117
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pandas_ta==0.3.14b0
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.8
propcache==0.3.1
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
pyparsing==3.2.0
pypdf==5.4.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stable-ts==2.15.11
starlette==0.46.2
striprtf==0.0.26
sympy==1.12
tenacity==9.1.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.2.2
tqdm==4.66.2
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.34.2
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.20.0
yfinance==0.2.50
zstandard==0.23.0
